feat: type MHA q/k-norm placeholders; full gemma3 traces weight-free
- MultiHeadAttention threads its `dtype` into the internal q_norm/k_norm
  RMSNormalization (qkNorm path, used by gemma3) — the last untyped placeholder.
- settings: opt-in composite build (-PuseLocalSkainet=true) to validate against
  a sibling ../SKaiNET checkout instead of Maven artifacts.
- GemmaTraceTest now COMPLETES: tiny gemma3 traces to a 116-node ComputeGraph
  weight-free. Op histogram shows every op has a StableHLO converter EXCEPT
  `split` x2 (RoPE) — which needs multi-output converter-framework support.

Verified via composite build against SKaiNET feature/stablehlo-dtype (which
carries the VoidTensorOps.gather shape fix + DSL dtype fixes).

+++ llm-core/src/commonMain/kotlin/sk/ainet/lang/nn/transformer/MultiHeadAttention.kt
@@ -164,11 +164,11 @@ public class MultiHeadAttention<T : DType, V>(
 
     // Optional QK-Norm layers
     public val qNorm: RMSNormalization<T, V>? = if (qkNorm) {
-        RMSNormalization(intArrayOf(headDim), eps = qkNormEps, name = "$name.q_norm", unitOffset = qkNormUnitOffset)
+        RMSNormalization(intArrayOf(headDim), eps = qkNormEps, name = "$name.q_norm", unitOffset = qkNormUnitOffset, dtype = dtype)
     } else null
 
     public val kNorm: RMSNormalization<T, V>? = if (qkNorm) {
-        RMSNormalization(intArrayOf(headDim), eps = qkNormEps, name = "$name.k_norm", unitOffset = qkNormUnitOffset)
+        RMSNormalization(intArrayOf(headDim), eps = qkNormEps, name = "$name.k_norm", unitOffset = qkNormUnitOffset, dtype = dtype)
     } else null
 
     @Suppress("UNCHECKED_CAST")
